Initial commit of Bingo! Deck Generator program.
This project is for our very important client.  They kept talking about
the great job my intern did with their last project.  How did they find
out that I didn't personally code it?

We have just finished working through the requirements for this program,
which are documented on Canvas.  Now that the hard part is complete I
think that I'll hand this job off to that intern again; it'll look good
on their resume and gives me more time to play Commander Keen.